Dijkstra vitesse et distance

+++ be-graphes-model/src/main/java/org/insa/graphs/model/Label.java
@@ -3,24 +3,24 @@ package org.insa.graphs.model;
 public class Label implements Comparable<Label> {
 	public Node sommet;
 	public boolean marque;
-	private float cost;
+	private double cost;
 	public Arc pere;
-	public Label(Node sommet, boolean marque, float cost, Arc pere ) {
+	public Label(Node sommet, boolean marque, double cost, Arc pere ) {
 		this.sommet=sommet;
 		this.marque=marque;
 		this.cost=cost;
 		this.pere=pere;
 		
 	}
-	public float getCost() {
+	public double getCost() {
 		return this.cost;
 	}
-	public void setCost(float newCost) {
+	public void setCost(double newCost) {
 		this.cost=newCost;
 	}
 	
 	public int compareTo(Label other) {
-		return Float.compare(this.getCost(),other.getCost());
+		return Double.compare(this.getCost(),other.getCost());
 	}
 
 }
